requirements.txt
django>=2.2.0,<3
djangorestframework<3.13,>=3.8.2
uwsgi==2.0.20
requests==2.27.1
pandas==1.4.0
pygdal==3.0.4.10
future==0.18.2
numpy==1.22.2
six==1.16.0
matplotlib==3.2.2
scipy==1.7.1
nose==1.3.7
cython==0.29.24
netcdf4==1.5.7
dill==0.3.4
future==0.18.2
pymetis==2020.1
triangle==20200424
Pmw==2.0.1
mpi4py==3.1.1
gitpython==3.1.24
boto3==1.17.97
#pyproj==
